renamed generic_functions to common
