+++ iSkyLIMS_wetlab/utils/handling_crontab_run_metric.py
@@ -6,7 +6,7 @@ from  iSkyLIMS_wetlab.models import RunProcess, RunningParameters, StatsRunSumma
 from django.conf import settings
 from iSkyLIMS_wetlab.wetlab_config import *
 
-from .generic_functions import logging_errors, logging_warnings
+from .common import logging_errors, logging_warnings
 from .handling_crontab_common_functions import *
